Adds more unit tests.
- Removes duplicated ArrayMergeRecursiveDistinctTest since this is
  covered in the Handler unit tests.
